Refactor class names to follow coding standards
Updated class names to align with Clean Code principles and Java
naming conventions. Classes now use nouns instead of verbs, making
their purpose clearer and improving maintainability.

Examples of changes made:
- Renamed `CheckDeadline` to `DeadlineValidator`
- Renamed `CheckEvent` to `EventValidator`
- Renamed `CheckMark` to `MarkValidator`
- Renamed `CheckUnmark` to `UnmarkValidator`
- Renamed `CheckDeadlineDuplicates` to `DuplicateDeadlineChecker`
- Renamed `CheckEventDuplicates` to `DuplicateEventChecker`
- Renamed `CheckTodoDuplicates` to `DuplicateTodoChecker`

These changes enhance code readability and ensure consistency
across the project.
