+++ src/main/java/chatbot/check/MarkValidator.java
@@ -1,12 +1,12 @@
 package chatbot.check;
 
-import chatbot.exceptions.UnmarkException;
+import chatbot.exceptions.MarkException;
 import chatbot.tasks.TaskList;
 
 /**
- * Utility class to validate the unmark command input.
+ * Utility class to validate the mark command input.
  */
-public class CheckUnmark {
+public class MarkValidator {
 
     /**
      * Validates if the given input is a numeric task index and within the valid range.
@@ -14,9 +14,9 @@ public class CheckUnmark {
      * @param input The raw input string representing a task index.
      * @param tasks The task list to check the index against.
      * @return The valid integer index.
-     * @throws UnmarkException If the input is not numeric or the index is out of range.
+     * @throws MarkException If the input is not numeric or the index is out of range.
      */
-    public static int validate(String input, TaskList tasks) throws UnmarkException {
+    public static int validate(String input, TaskList tasks) throws MarkException {
         assert input != null : "Input task number cannot be null";
 
         int index;
@@ -25,14 +25,15 @@ public class CheckUnmark {
         try {
             index = Integer.parseInt(input);
         } catch (NumberFormatException e) {
-            throw new UnmarkException("OOPS!!! The unmark command requires a valid numeric task number.");
+            throw new MarkException("OOPS!!! The mark command requires a valid numeric task number.");
         }
 
         // Check if the index is within range
         if (index < 1 || index > tasks.size()) {
-            throw new UnmarkException("OOPS!!! The task number provided is out of range.");
+            throw new MarkException("OOPS!!! The task number provided is out of range.");
         }
 
         return index;
     }
 }
+
