+++ src/main/java/chatbot/check/UnmarkValidator.java
@@ -1,12 +1,12 @@
 package chatbot.check;
 
-import chatbot.exceptions.MarkException;
+import chatbot.exceptions.UnmarkException;
 import chatbot.tasks.TaskList;
 
 /**
- * Utility class to validate the mark command input.
+ * Utility class to validate the unmark command input.
  */
-public class CheckMark {
+public class UnmarkValidator {
 
     /**
      * Validates if the given input is a numeric task index and within the valid range.
@@ -14,9 +14,9 @@ public class CheckMark {
      * @param input The raw input string representing a task index.
      * @param tasks The task list to check the index against.
      * @return The valid integer index.
-     * @throws MarkException If the input is not numeric or the index is out of range.
+     * @throws UnmarkException If the input is not numeric or the index is out of range.
      */
-    public static int validate(String input, TaskList tasks) throws MarkException {
+    public static int validate(String input, TaskList tasks) throws UnmarkException {
         assert input != null : "Input task number cannot be null";
 
         int index;
@@ -25,15 +25,14 @@ public class CheckMark {
         try {
             index = Integer.parseInt(input);
         } catch (NumberFormatException e) {
-            throw new MarkException("OOPS!!! The mark command requires a valid numeric task number.");
+            throw new UnmarkException("OOPS!!! The unmark command requires a valid numeric task number.");
         }
 
         // Check if the index is within range
         if (index < 1 || index > tasks.size()) {
-            throw new MarkException("OOPS!!! The task number provided is out of range.");
+            throw new UnmarkException("OOPS!!! The task number provided is out of range.");
         }
 
         return index;
     }
 }
-
